+++ packages/block-library/src/embed/settings.js
@@ -76,7 +76,7 @@ export function getEmbedBlockSettings( { title, description, icon, category = 'e
 				return {
 					preview: validPreview ? preview : undefined,
 					fetching,
-					supportsResponsive: themeSupports[ 'responsive-embeds' ],
+					themeSupportsResponsive: themeSupports[ 'responsive-embeds' ],
 					cannotEmbed,
 				};
 			} )